python/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
cachecontrol==0.12.6
    # via firebase-admin
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via google-crc32c
charset-normalizer==2.0.6
    # via requests
click==8.0.1
    # via flask
colorama==0.4.4
    # via click
firebase-admin==5.0.2
    # via -r requirements.in
flask==2.0.1
    # via -r requirements.in
google-api-core[grpc]==1.31.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.21.0
    # via firebase-admin
google-auth==1.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.0.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.3.2
    # via firebase-admin
google-cloud-storage==1.42.2
    # via firebase-admin
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==2.0.2
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
grpcio==1.40.0
    # via google-api-core
httplib2==0.19.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.2
    # via requests
imagehash==4.2.1
    # via -r requirements.in
imutils==0.5.4
    # via -r requirements.in
itsdangerous==2.0.1
    # via flask
jinja2==3.0.1
    # via flask
markupsafe==2.0.1
    # via jinja2
msgpack==1.0.2
    # via cachecontrol
numpy==1.21.2
    # via
    #   -r requirements.in
    #   imagehash
    #   opencv-python
    #   pywavelets
    #   scipy
opencv-python==4.5.3.56
    # via -r requirements.in
packaging==21.0
    # via
    #   google-api-core
    #   google-cloud-firestore
pillow==8.3.2
    # via imagehash
proto-plus==1.19.0
    # via google-cloud-firestore
protobuf==3.18.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
python-dotenv==0.19.0
    # via -r requirements.in
pytz==2021.1
    # via google-api-core
pywavelets==1.1.1
    # via imagehash
requests==2.26.0
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
    #   requests-auth
requests-auth==5.3.0
    # via -r requirements.in
rsa==4.7.2
    # via google-auth
scipy==1.7.1
    # via imagehash
six==1.16.0
    # via
    #   google-api-core
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-storage
    #   grpcio
    #   imagehash
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.6
    # via requests
werkzeug==2.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
